requirements.txt
gdown==5.2.0
logger==1.4
matplotlib==3.7.1
numpy==1.24.2
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.3.1
scipy==1.14.1
timm==0.9.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.37.2
